requirements.txt
pdfminer.six==20231228
streamlit==1.36.0
